KicomBasic/requirements.txt
pip==19.0.3
pycryptodome==3.8.0
setuptools==40.8.0
